+++ src/lib/service.ts
@@ -74,6 +74,10 @@ export async function isBatchMkvExtractFound(path: string, checkRunning: boolean
   return await invoke<Protocol.BatchMkvExtractStatus>("is_batchmkvextract_found", { path, checkRunning });
 }
 
+export async function openBatchMkvExtract(file: string): Promise<void> {
+  return await invoke<void>("open_batchmkvextract", { file });
+}
+
 export async function runMkvextract(file: string, args: string[]): Promise<void> {
   return await invoke<void>("run_mkvextract", { file, args });
 }